web/pyproject.toml
[tool.poetry]
name = "stkweb"
version = "0.1.0"
description = ""
authors = ["John Anderson <sontek@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
gunicorn = "^20.0.4"
uvicorn = "^0.13.3"
fastapi = "^0.63.0"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
